fix gamerules for controlling block-damage and fire-damage from explosions
