+++ src/main/java/dev/amble/ait/core/blocks/PeanutBlock.java
@@ -1,10 +1,12 @@
 package dev.amble.ait.core.blocks;
 
+import dev.amble.ait.core.tardis.util.TardisUtil;
 import net.minecraft.block.Block;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
 public class PeanutBlock extends Block {
+    private static final float EXPLOSION_POWER = 100;
 
     public PeanutBlock(Settings settings) {
         super(settings.strength(-1.0f, 3600000.0f).emissiveLighting((state, world, pos) -> true).luminance(value -> 128)
@@ -12,10 +14,10 @@ public class PeanutBlock extends Block {
     }
 
     public void explode(World world, BlockPos pos) {
-        world.createExplosion(null, world.getDamageSources().outOfWorld(), null, pos.toCenterPos(), 100, true,
+        world.createExplosion(null, world.getDamageSources().outOfWorld(), TardisUtil.EXPLOSION_BEHAVIOR, pos.toCenterPos(), EXPLOSION_POWER, TardisUtil.doCreateFire(world),
                 World.ExplosionSourceType.MOB);
-        world.createExplosion(null, null, null, pos.toCenterPos(), 100, true, World.ExplosionSourceType.BLOCK);
-        world.createExplosion(null, world.getDamageSources().outOfWorld(), null, pos.toCenterPos(), 100, true,
+        world.createExplosion(null, null, TardisUtil.EXPLOSION_BEHAVIOR, pos.toCenterPos(), EXPLOSION_POWER, TardisUtil.doCreateFire(world), World.ExplosionSourceType.BLOCK);
+        world.createExplosion(null, world.getDamageSources().outOfWorld(), TardisUtil.EXPLOSION_BEHAVIOR, pos.toCenterPos(), EXPLOSION_POWER, TardisUtil.doCreateFire(world),
                 World.ExplosionSourceType.TNT);
     }
 }
